Java: Generate filtered dao from modeled version.
